requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.0
fonttools==4.58.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.5
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
setuptools==80.4.0
six==1.17.0
Werkzeug==3.1.3
